Falling square v2.0
I don't know how, but I made it to work with the actual gravity (with a couple of questionable lines of code). It has some bugs, specially when the grid is at least ~70% filled with squares, they start to behave a little bit weird (they start to move the the left a bit too much).